requirements.txt
django==5.0.6
celery==5.4.0
redis==5.0.1
fastapi==0.111.0
uvicorn==0.30.0
requests==2.32.3
django-tinymce==3.5.0
psycopg2-binary==2.9.9
pytest==8.2.0